Enexia/App/requirements.txt
streamlit>=1.32.0
groq>=0.5.0
python-dotenv>=1.0.0
python-frontmatter>=1.1.0
watchdog>=4.0.0
